fix dummy_oracle::evaluated_contents
This function reserves some number of entries in the result vector
with the intent of adding that many entries. Instead, the for loop
iterates over the *size* of results (which is zero---reserve doesn't
change the size'!').

This causes evaluated_contents to return a zero-length vector, which
eventually bubbles up and causes a segfault in fri_iop.